+++ packages/next/src/utilities/getNextI18n.ts
@@ -6,13 +6,13 @@ import { cookies, headers } from 'next/headers.js'
 
 import { getRequestLanguage } from './getRequestLanguage.js'
 
-export const getNextI18n = async ({
+export const getNextI18n = ({
   config,
   language,
 }: {
   config: SanitizedConfig
   language?: string
-}): Promise<I18n> =>
+}): I18n =>
   initI18n({
     config: config.i18n,
     context: 'client',
