+++ packages/next/src/views/NotFound/index.tsx
@@ -19,7 +19,7 @@ export const generatePageMetadata = async ({
 }): Promise<Metadata> => {
   const config = await configPromise
 
-  const i18n = await getNextI18n({
+  const i18n = getNextI18n({
     config,
   })
